Move cast test to InlineCastTest
