+++ drools-model/drools-model-compiler/src/test/java/org/drools/modelcompiler/CompilerTest.java
@@ -2295,61 +2295,4 @@ public class CompilerTest extends BaseModelTest {
 
         Assertions.assertThat(list).containsExactlyInAnyOrder(40, 38, 41, 38, 41, 43);
     }
-
-
-    public interface DecisionTable extends Expression {
-        List<OutputClause> getOutput();
-    }
-
-    public interface OutputClause extends DMNElement {
-
-    }
-
-    public interface Expression extends DMNElement {
-
-    }
-
-    public interface DMNElement extends DMNModelInstrumentedBase {
-
-    }
-
-    public interface DMNModelInstrumentedBase {
-        DMNModelInstrumentedBase getParent();
-    }
-
-    @Test
-    public void testParent() {
-        String str =
-                "import " + OutputClause.class.getCanonicalName() + "\n;" +
-                "import " + DecisionTable.class.getCanonicalName() + "\n;" +
-                "rule r\n" +
-                "when\n" +
-                "  $oc : OutputClause( parent instanceof DecisionTable, parent.output.size > 1 )\n" +
-                "then\n" +
-                "end\n";
-
-        KieSession ksession = getKieSession( str );
-
-        ksession.fireAllRules();
-
-        assertTrue(true);
-    }
-
-    @Test
-    public void testParentInlineCast() {
-        String str =
-                "import " + OutputClause.class.getCanonicalName() + "\n;" +
-                        "import " + DecisionTable.class.getCanonicalName() + "\n;" +
-                        "rule r\n" +
-                        "when\n" +
-                        "  $oc : OutputClause( parent#DecisionTable.output.size > 1 )\n" +
-                        "then\n" +
-                        "end\n";
-
-        KieSession ksession = getKieSession( str );
-
-        ksession.fireAllRules();
-
-        assertTrue(true);
-    }
 }
